associated socket events in same file
